requirements.txt
scikit-learn
numpy
pandas
matplotlib
seaborn
flask
gunicorn
streamlit
pillow